Fixed a bunch of forecasts for NAC centers
A bunch of the NAC centers got new region IDs this year. I've updated
the region ids to match the current ones, but I didn't update any other
region information. We may be doing that as a separate task, but I
wanted to at least update the IDs so the map shows the warnings.