Automate updates to NAC and Canadian Regions
We use a regions.json file to hold metadata on all of the US and Canadian
regions, including polygon coordinates for mapping. Instead of spending
time out of sync and having to manually update this file, we can use US and
Canadian national data sources to keep this data current automatically.
This update standardizes the naming of regions, adds centerId to the
regions file for easier processing, adjusts logic that had been using old
regionIds to either use the new regionId or the centerId, and sets up a
daily pull of region data and regeneration of the regions file.

Issue #55533