Issue #57584: Updated CAIC Region IDs
Andy noticed that the Colorado polygons stopped updating this month even
though forecasts continue to be issued. The polygon and region data is now
automatically updated nightly, but we still have code that is not dynamic
and relies on knowing the region id in order to load a region's data. With
recent updates to CAIC centers, they received new ids, and this is the
necessary code update to match.