Issue #57843: Date handling for avalanche-canada regions
CAC regions have forecasts with a single date attached to them which we've
been treating as a start date. Based on the Issue Date and Valid Date from
the avalanche.ca website, this is more accurate as an end date, meaning
that we should subtract a day to have a start date. In addition to this, we
still want to copy the first forecast and backfill a day to cover cases of
a forecast being published several days in advance so that we aren't left
with an empty day.